DTOs

+++ store/src/main/java/com/ytamang/store/entities/Product.java
@@ -22,7 +22,7 @@ public class Product {
     @Id
     @Column(name = "id")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Byte id;
+    private Long id;
 
     @Column(name = "name")
     private String name;
